Modulo 4

+++ src/main/java/com/jackson/jfood/infrastructure/repository/CityRepositoryImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -35,8 +36,10 @@ public class CityRepositoryImpl implements CityRepository {
 
 	@Transactional
 	@Override
-	public void remove(City city) {
-		city = find(city.getId());
+	public void remove(Long id) {
+		City city = find(id);
+		if (city == null)
+			throw new EmptyResultDataAccessException(1);
 		entityManager.remove(city);
 	}
