+++ src/main/java/com/jackson/jfood/infrastructure/repository/CuisineRepositoryImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -35,8 +36,10 @@ public class CuisineRepositoryImpl implements CuisineRepository {
 	
 	@Transactional
 	@Override
-	public void remove(Cuisine cuisine) {
-		cuisine = find(cuisine.getId());
+	public void remove(Long id) {
+		Cuisine cuisine = find(id);
+		if (cuisine == null)
+			throw new EmptyResultDataAccessException(1);
 		entityManager.remove(cuisine);
 	}
