+++ src/main/java/com/jackson/jfood/infrastructure/repository/StateRepositoryImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -35,8 +36,10 @@ public class StateRepositoryImpl implements StateRepository {
 
 	@Transactional
 	@Override
-	public void remove(State state) {
-		state = find(state.getId());
+	public void remove(Long id) {
+		State state = find(id);
+		if (state == null)
+			throw new EmptyResultDataAccessException(1);
 		entityManager.remove(state);
 	}
